+++ nginx-core/etc/nginx/entrypoint.d/93-cors.sh
@@ -1,10 +1,10 @@
 #!/bin/sh
 set -eu
 
-ME=$(basename "$0")
-
+test -n "${TEST_MODE:-}" && exit 0
 test -n "${NGINX_CORS_ENABLE:-}" && exit 0
 
+ME=$(basename "$0")
 echo "$ME: Remove CORS config"
 rm -fv /etc/nginx/conf.d/location.d/70-cors.conf
